+++ publish-service/src/test/java/com/ericsson/eiffel/remrem/publish/service/EiffelRemremCommonControllerUnitTest.java
@@ -126,7 +126,7 @@ public class EiffelRemremCommonControllerUnitTest {
         when(restTemplate.postForEntity(Mockito.contains(correctURL), Mockito.<HttpEntity<String>>any(),
                 Mockito.eq(String.class), Mockito.anyMap())).thenReturn(responseOK);
 
-        ResponseEntity<?> elem = unit.generateAndPublish("eiffelsemantics", "eiffelactivityfinished", "", "", "",false,
+        ResponseEntity<?> elem = unit.generateAndPublish("eiffelsemantics", "eiffelactivityfinished", null, "", "", "",false,
                 null, null, true, 1, false, body.getAsJsonObject());
         assertEquals(elem.getStatusCode(), HttpStatus.OK);
 
@@ -140,7 +140,7 @@ public class EiffelRemremCommonControllerUnitTest {
         when(restTemplate.postForEntity(Mockito.contains(inCorrectURL), Mockito.<HttpEntity<String>>any(),
                 Mockito.eq(String.class), Mockito.anyMap())).thenReturn(responseBad);
 
-        ResponseEntity<?> elem = unit.generateAndPublish("eiffel3", "eiffelactivityfinished", "", "", "",false,
+        ResponseEntity<?> elem = unit.generateAndPublish("eiffel3", "eiffelactivityfinished", null, "", "", "",false,
                 null, null, true, 1, false, body.getAsJsonObject());
         assertEquals(elem.getStatusCode(), HttpStatus.BAD_REQUEST);
 
@@ -176,7 +176,7 @@ public class EiffelRemremCommonControllerUnitTest {
         when(restTemplate.postForEntity(Mockito.contains(correctURL), Mockito.<HttpEntity<String>> any(),
                 Mockito.eq(String.class), Mockito.anyMap())).thenReturn(responseOptionsFailed);
 
-        ResponseEntity<?> elem = unit.generateAndPublish("eiffelsemantics", "eiffelactivityfinished", "", "", "", false,
+        ResponseEntity<?> elem = unit.generateAndPublish("eiffelsemantics", "eiffelactivityfinished", null, "", "", "", false,
                 false, false, true, 1, false, body.getAsJsonObject());
         assertEquals(elem.getStatusCode(), HttpStatus.UNPROCESSABLE_ENTITY);
 
@@ -189,7 +189,7 @@ public class EiffelRemremCommonControllerUnitTest {
         when(restTemplate.postForEntity(Mockito.contains(correctURL), Mockito.<HttpEntity<String>> any(),
                 Mockito.eq(String.class), Mockito.anyMap())).thenReturn(responseMultipleFound);
 
-        ResponseEntity<?> elem = unit.generateAndPublish("eiffelsemantics", "eiffelactivityfinished", "", "", "", false,
+        ResponseEntity<?> elem = unit.generateAndPublish("eiffelsemantics", "eiffelactivityfinished", null, "", "", "", false,
                 false, false, true, 1, false, body.getAsJsonObject());
         assertEquals(elem.getStatusCode(), HttpStatus.EXPECTATION_FAILED);
 
@@ -202,7 +202,7 @@ public class EiffelRemremCommonControllerUnitTest {
         when(restTemplate.postForEntity(Mockito.contains(correctURL), Mockito.<HttpEntity<String>> any(),
                 Mockito.eq(String.class), Mockito.anyMap())).thenReturn(responseNoneFound);
 
-        ResponseEntity<?> elem = unit.generateAndPublish("eiffelsemantics", "eiffelactivityfinished", "", "", "", false,
+        ResponseEntity<?> elem = unit.generateAndPublish("eiffelsemantics", "eiffelactivityfinished", null, "", "", "", false,
                 false, false, true, 1, false, body.getAsJsonObject());
         assertEquals(elem.getStatusCode(), HttpStatus.NOT_ACCEPTABLE);