+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Implemented the changes to log the eventId and HTTPStatus while the level is INFO.
 - Implemented the changes to print the user information while the log level is INFO.
 - Made changes to resolve extra RabbitMQ connection issue.
+- Uplifted the amqp-client version to 5.20.0
 
 ## 2.1.0
 - Implemented new routing key template for Sepia.